fix: populate fixed and API link

+++ backend/controllers/populate.js
@@ -14,7 +14,6 @@ const {
     GlobalConfiguration,
     Suggestion
 } = require("../models");
-
 const fs = require('fs');
 const path = require("path");
 const fullDataPath = path.resolve("backend/populate.json");
@@ -43,18 +42,25 @@ var create_report = (model_name, result, data) => {
 };
 
 exports.populate = async(req, res, next) => {
+    console.log("Populating database...")
+    console.log("Using host: " + req.get("host"))
     var rawData = null;
     if (req.get("host").includes("localhost")) {
         rawData = fs.readFileSync(fullDataPath);
     } else {
         rawData = fs.readFileSync(fullProdDataPath);
-        console.log(rawData)
     }
     const populate_json_data = JSON.parse(rawData);
     reports = { "errored_models": [], "reports": [] };
+
     //users
     try {
-        User.collection.drop();
+        console.log("Users:")
+        console.log(await User.exists({}))
+        if (await User.exists({})) {
+            await User.collection.drop();
+        }
+        await User.createCollection();
         var usersResult = await User.insertMany(populate_json_data.users, { ordered: false, rawResult: true });
         [message, rep] = create_report('user', usersResult, populate_json_data.users);
         reports.reports.push({ "message": message, "report": rep });
@@ -63,7 +69,12 @@ exports.populate = async(req, res, next) => {
     }
     // workspaces
     try {
-        Workspace.collection.drop();
+        console.log("Workspaces:")
+        console.log(await Workspace.exists({}))
+        if (await Workspace.exists({})) {
+            await Workspace.collection.drop();
+        }
+        await Workspace.createCollection();
         var workspacesResult = await Workspace.insertMany(populate_json_data.workspaces, { ordered: false, rawResult: true });
         [message, rep] = create_report('workspace', workspacesResult, populate_json_data.workspaces);
         reports.reports.push({ "message": message, "report": rep });
@@ -72,7 +83,12 @@ exports.populate = async(req, res, next) => {
     }
     // roles
     try {
-        Role.collection.drop();
+        console.log("Roles:")
+        console.log(await Role.exists({}))
+        if (await Role.exists({})) {
+            await Role.collection.drop();
+        }
+        await Role.createCollection();
         var rolesResult = await Role.insertMany(populate_json_data.roles, { ordered: false, rawResult: true });
         [message, rep] = create_report('role', rolesResult, populate_json_data.roles);
         reports.reports.push({ "message": message, "report": rep });
@@ -82,7 +98,12 @@ exports.populate = async(req, res, next) => {
 
     // invitations
     try {
-        Invitation.collection.drop();
+        console.log("Invitations:")
+        console.log(await Invitation.exists({}))
+        if (await Invitation.exists({})) {
+            await Invitation.collection.drop();
+        }
+        await Invitation.createCollection();
         var invitationsResult = await Invitation.insertMany(populate_json_data.invitations, { ordered: false, rawResult: true });
         [message, rep] = create_report('invitation', invitationsResult, populate_json_data.invitations);
         reports.reports.push({ "message": message, "report": rep });
@@ -92,7 +113,12 @@ exports.populate = async(req, res, next) => {
 
     // Collections
     try {
-        Collection.collection.drop();
+        console.log("Collections:")
+        console.log(await Collection.exists({}))
+        if (await Collection.exists({})) {
+            await Collection.collection.drop();
+        }
+        await Collection.createCollection();
         var collectionsResult = await Collection.insertMany(populate_json_data.collections, { ordered: false, rawResult: true });
         [message, rep] = create_report('collection', collectionsResult, populate_json_data.collections);
         reports.reports.push({ "message": message, "report": rep });
@@ -102,7 +128,12 @@ exports.populate = async(req, res, next) => {
 
     // Datafiles
     try {
-        Datafile.collection.drop();
+        console.log("Datafiles:")
+        console.log(await Datafile.exists({}))
+        if (await Datafile.exists({})) {
+            await Datafile.collection.drop();
+        }
+        await Datafile.createCollection();
         var datafilesResult = await Datafile.insertMany(populate_json_data.datafiles, { ordered: false, rawResult: true });
         [message, rep] = create_report('datafile', datafilesResult, populate_json_data.datafiles);
         reports.reports.push({ "message": message, "report": rep });
@@ -112,7 +143,12 @@ exports.populate = async(req, res, next) => {
 
     // Esquemas
     try {
-        Esquema.collection.drop();
+        console.log("Esquemas:")
+        console.log(await Esquema.exists({}))
+        if (await Esquema.exists({})) {
+            await Esquema.collection.drop();
+        }
+        await Esquema.createCollection();
         var esquemasResult = await Esquema.insertMany(populate_json_data.esquemas, { ordered: false, rawResult: true });
         [message, rep] = create_report('esquema', esquemasResult, populate_json_data.esquemas);
         reports.reports.push({ "message": message, "report": rep });
@@ -121,7 +157,12 @@ exports.populate = async(req, res, next) => {
     }
 
     try {
-        Configuration.collection.drop();
+        console.log("Configurations:")
+        console.log(await Configuration.exists({}))
+        if (await Configuration.exists({})) {
+            await Configuration.collection.drop();
+        }
+        await Configuration.createCollection();
         var configurationsResult = await Configuration.insertMany(populate_json_data.configurations, { ordered: false, rawResult: true });
         [message, rep] = create_report('configuration', configurationsResult, populate_json_data.configurations);
         reports.reports.push({ "message": message, "report": rep });
@@ -131,7 +172,12 @@ exports.populate = async(req, res, next) => {
 
     // Tests
     try {
-        Test.collection.drop();
+        console.log("Tests:")
+        console.log(await Test.exists({}))
+        if (await Test.exists({})) {
+            await Test.collection.drop();
+        }
+        await Test.createCollection();
         var testsResult = await Test.insertMany(populate_json_data.tests, { ordered: false, rawResult: true });
         [message, rep] = create_report('test', testsResult, populate_json_data.tests);
         reports.reports.push({ "message": message, "report": rep });
@@ -141,7 +187,12 @@ exports.populate = async(req, res, next) => {
 
     // Activities
     try {
-        Activity.collection.drop();
+        console.log("Activities:")
+        console.log(await Activity.exists({}))
+        if (await Activity.exists({})) {
+            await Activity.collection.drop();
+        }
+        await Activity.createCollection();
         var activitiesResult = await Activity.insertMany(populate_json_data.activities, { ordered: false, rawResult: true });
         [message, rep] = create_report('activity', activitiesResult, populate_json_data.activities);
         reports.reports.push({ "message": message, "report": rep });
@@ -151,7 +202,12 @@ exports.populate = async(req, res, next) => {
 
     // Terminals
     try {
-        Terminal.collection.drop();
+        console.log("Terminals:")
+        console.log(await Terminal.exists({}))
+        if (await Terminal.exists({})) {
+            await Terminal.collection.drop();
+        }
+        await Terminal.createCollection();
         var terminalsResult = await Terminal.insertMany(populate_json_data.terminals, { ordered: false, rawResult: true });
         [message, rep] = create_report('terminal', terminalsResult, populate_json_data.terminals);
         reports.reports.push({ "message": message, "report": rep });
@@ -161,7 +217,12 @@ exports.populate = async(req, res, next) => {
 
     // FricErrors
     try {
-        FricError.collection.drop();
+        console.log("FricErrors:")
+        console.log(await FricError.exists({}))
+        if (await FricError.exists({})) {
+            await FricError.collection.drop();
+        }
+        await FricError.createCollection();
         var fricErrorsResult = await FricError.insertMany(populate_json_data.fricErrors, { ordered: false, rawResult: true });
         [message, rep] = create_report('fricError', fricErrorsResult, populate_json_data.fricErrors);
         reports.reports.push({ "message": message, "report": rep });
@@ -171,7 +232,12 @@ exports.populate = async(req, res, next) => {
 
     // Global Config
     try {
-        GlobalConfiguration.collection.drop();
+        console.log("Global Configuration:")
+        console.log(await GlobalConfiguration.exists({}))
+        if (await GlobalConfiguration.exists({})) {
+            await GlobalConfiguration.collection.drop();
+        }
+        await GlobalConfiguration.createCollection();
         var globalConfigResult = await GlobalConfiguration.insertMany(populate_json_data.globalConfiguration, { ordered: false, rawResult: true });
         [message, rep] = create_report('globalConfig', globalConfigResult, populate_json_data.globalConfiguration);
         reports.reports.push({ "message": message, "report": rep });
@@ -180,14 +246,19 @@ exports.populate = async(req, res, next) => {
     }
 
     // Suggestions
-    // try {
-    //     Suggestion.collection.drop();
-    //     var suggestionsResult = await Suggestion.insertMany(populate_json_data.suggestions, { ordered: false, rawResult: true });
-    //     [message, rep] = create_report('suggestion', suggestionsResult, populate_json_data.suggestions);
-    //     reports.reports.push({ "message": message, "report": rep });
-    // } catch (err) {
-    //     reports.errored_models.push({ "Suggestion": err });
-    // }
+    try {
+        console.log("Suggestions:")
+        console.log(await Suggestion.exists({}))
+        if (await Suggestion.exists({})) {
+            await Suggestion.collection.drop();
+        }
+        await Suggestion.createCollection();
+        var suggestionsResult = await Suggestion.insertMany(populate_json_data.suggestions, { ordered: false, rawResult: true });
+        [message, rep] = create_report('suggestion', suggestionsResult, populate_json_data.suggestions);
+        reports.reports.push({ "message": message, "report": rep });
+    } catch (err) {
+        reports.errored_models.push({ "Suggestion": err });
+    }
 
     fs.readdir(path.join("backend/uploads/datafiles"), (err, files) => {
         if (err) console.log(err);
@@ -218,7 +289,7 @@ exports.populate = async(req, res, next) => {
             });
         }
     });
-
+    console.log("Fin")
     return res.status(200).json({
         data: reports
     });
